requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
zipp==3.10.0
